added concurrently to be able to run frontend and backend with one command, also added a confirmation box for deleting a player

+++ View/frontend/src/components/deletebutton.jsx
@@ -1,6 +1,16 @@
 function DeleteButton({ playerId, fetchPlayers }) {
 
     async function handleDelete() {
+
+
+        const confirmed = window.confirm(
+            "Are you sure you want to delete this player?"
+        );
+
+        if (!confirmed) {
+            return;
+        }
+
         const response = await fetch(`http://localhost:3000/api/players/${playerId}`, {
             method: "DELETE"
         });
